+++ pkg/receiver/jobreceiver/asset/environment_windows.go
@@ -1,0 +1,13 @@
+//go:build windows
+// +build windows
+
+package asset
+
+import "strings"
+
+// On Windows, environment variables are case-insensitive; to avoid conflicts we
+// coerce all keys to UPPER CASE.
+// https://docs.microsoft.com/en-us/dotnet/api/system.environment.getenvironmentvariable?view=netframework-4.7.2
+func coerceKey(k string) string {
+	return strings.ToUpper(k)
+}